final version

+++ trafficmon/src/com/trafficmon/CustomerAccountsService.java
@@ -4,6 +4,8 @@ package com.trafficmon;
     This is adaptor interface for handling payment and customer accounts issues which supplied by third-party library.
  */
 
+import java.math.BigDecimal;
+
 public interface CustomerAccountsService {
-    public abstract void deductCharge(Vehicle vehicle, long charge) throws InsufficientCreditException, AccountNotRegisteredException;
+    public abstract void deductCharge(Vehicle vehicle, BigDecimal charge) throws InsufficientCreditException, AccountNotRegisteredException;
 }
