+++ trafficmon/src/com/trafficmon/OperationService.java
@@ -8,5 +8,5 @@ import java.math.BigDecimal;
 
 public interface OperationService {
     public abstract void investigate(Vehicle vehicle);
-    public abstract void penaltyNotice(Vehicle vehicle, long charge);
+    public abstract void penaltyNotice(Vehicle vehicle, BigDecimal charge);
 }
